Using mutate.appendChild so that <select> is set correctly
In IE9/IE10, <select>'s selectedIndex is set incorrectly when inserted
into the document. This is corrected in can-util by setting the selectedIndex
when the "inserted" event fires. This change makes it so the test uses
mutat.appendChild in order to trigger "inserted" events in browsers that do
not support MutationObserver.