making on:click="does().some().thing()" work
This works by calling the expression directly instead of trying to look
up the value in the scope. It is wrapped in a queue for debuggability.
It also automatically calls a function that is returned by `thing()` in
order to preserve backward compatibility with functions that need to
implicitly get called with the element (for animations and things like
that).